+++ Cpoolday06/my_strncpy.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-int get_length(char const *str)
+size_t get_length(char const *str)
 {
-  int i = 0;
+  size_t i = 0;
 
   for (; str[i] != '\0'; ++i);
   return (i);
@@ -12,8 +12,8 @@ int get_length(char const *str)
 
 char *my_strncpy(char *dest, char const *src, int n)
 {
-  int i = 0;
-  int length = get_length(src);
+  size_t i = 0;
+  size_t length = get_length(src);
 
   if (n > length)
     dest[length] = '\0';
